requirements.txt
pytest
requests~=2.25.1
six~=1.15.0
boto3~=1.16.39
cachetools~=4.2.0
setuptools~=51.0.0
pycryptodome
pytz~=2020.5
confuse~=1.4.0
ratelimit~=2.2.1